Finished making Player and Gameboard compatible
Altered Player interface to use PlayerType enum
Added local PlayerType enum to ComputerPlayer and HumanPlayer classes,
which is set in the constructor for each class.

ComputerPlayer should now use getComputerFleet() to transfer ships (with
locations set) to the calling program. Calling program can then use
getStartTile and getEndTile methods, which I added to the Ship class, to
get the start and end tile for ship. These numbers can then be passed to
Gameboard through the deploy() method, which in addition to the start
and end numbers takes the player's PlayerType as a parameter.